Rename main.cpp to Cam_mov_glulookat.cpp
glulookat() function easy to implement but as per my logic there is one drawback. The drawback of using this function is that when move up on the top (of object at which you are looking), the object gets flipped 180 deg. For 3D modeling Software this is not good thing, it may disturb the flow flow of user while moving up.